requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
feedparser==6.0.11
APScheduler==3.10.4
requests==2.32.3
googletrans==4.0.0rc1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
neo4j==5.26.0
networkx==3.4.2
tweepy==4.14.0
